magit-ediff-stage: use same encoding for all texts
When using ediff to stage, the working tree file is visited by Emacs in
the normal way, so Emacs applies normal encoding detection on it.  We
must apply the same encoding to the output of `git cat-file' (used by
magit-find-file-{,index}-noselect) so that ediff can apply diffs
sensibly.

When writing back the changes back to git (via magit-update-index), we
have to use the same encoding to write as the one used to read, or else
we may end up changing the file recorded in git in a completely
unintended way.